+++ CHANGES.md
@@ -10,6 +10,7 @@ Change Log
 * Fixed default 3d-tiles styling to add a workaround for a Cesium bug which resulted in wrong translucency value for point clouds.
 * Remove Pell dependency, now replaced with TinyMCE (WYSIWYG editor library).
 * Fix `includes` path for `url-loader` rule so that it doesn't incorrectly match package names with `terriajs` as prefix.
+* `ChartPanelDownloadButton` will now only export columns which are visible in chart
 * [The next improvement]
 
 #### 8.2.5 - 2022-06-07